decouple `ShardKeySelector`, add shard key param
