+++ lib/collection/src/operations/mod.rs
@@ -7,7 +7,6 @@ pub mod operation_effect;
 pub mod payload_ops;
 pub mod point_ops;
 pub mod query_enum;
-pub mod shard_key_selector;
 pub mod shard_selector_internal;
 pub mod shared_storage_config;
 pub mod snapshot_ops;
