--- lib/collection/src/operations/shard_key_selector.rs
@@ -1,11 +1,0 @@
-use schemars::JsonSchema;
-use segment::types::ShardKey;
-use serde::{Deserialize, Serialize};
-
-#[derive(Debug, Deserialize, Serialize, Clone, JsonSchema, PartialEq)]
-#[serde(untagged)]
-pub enum ShardKeySelector {
-    ShardKey(ShardKey),
-    ShardKeys(Vec<ShardKey>),
-    // ToDo: select by pattern
-}
